Castor3D: Moved VCT related classes to GlobalIllumination folder
