+++ include/Core/Castor3D/Render/GlobalIllumination/VoxelConeTracing/VoxelSceneData.hpp
