+++ source/Core/Castor3D/Render/GlobalIllumination/VoxelConeTracing/VoxelizePass.cpp
@@ -1,4 +1,4 @@
-#include "Castor3D/Render/Technique/Voxelize/VoxelizePass.hpp"
+#include "Castor3D/Render/GlobalIllumination/VoxelConeTracing/VoxelizePass.hpp"
 
 #include "Castor3D/Engine.hpp"
 #include "Castor3D/Buffer/UniformBuffer.hpp"
@@ -14,7 +14,7 @@
 #include "Castor3D/Render/Culling/SceneCuller.hpp"
 #include "Castor3D/Render/Node/QueueCulledRenderNodes.hpp"
 #include "Castor3D/Render/Technique/RenderTechniqueVisitor.hpp"
-#include "Castor3D/Render/Technique/Voxelize/VoxelSceneData.hpp"
+#include "Castor3D/Render/GlobalIllumination/VoxelConeTracing/VoxelSceneData.hpp"
 #include "Castor3D/Scene/Camera.hpp"
 #include "Castor3D/Scene/Scene.hpp"
 #include "Castor3D/Scene/SceneNode.hpp"
